requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
llama-cpp-python>=0.2.0
pydantic>=2.0.0
jinja2>=3.0.0
markdown>=3.4.0
python-dotenv>=1.0.0
requests>=2.30.0
tqdm>=4.64.0
click>=8.0.0
colorama>=0.4.6
xmltodict>=0.13.0
click-plugins>=1.1.1
XlsxWriter>=3.0.0
numpy<2.0.0
huggingface_hub>=0.21.0
joblib>=1.3.0
transformers>=4.36.0,<4.52.0
torch>=2.0.0
accelerate>=0.27.0
bitsandbytes>=0.41.0

# Code analysis
tree-sitter==0.20.1
tree_sitter_languages==1.8.0
pygments>=2.15.0
semgrep>=1.30.0

# Machine learning & embedding
sentence-transformers>=2.2.0
faiss-cpu==1.7.4
openai>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Visualization
matplotlib>=3.7.0
networkx>=2.8.0
graphviz>=0.20.0 

# Optional dependencies
# Uncomment to install:

# For remote repository analysis (via Git)
# dulwich>=0.21.6 
